+++ src/module.js
@@ -16,17 +16,15 @@ var dependencies = [
   'ngSanitize', 'ui.router', 'restangular', 'satellizer',
   'ui.bootstrap.modal', 'ui.bootstrap.collapse', 'schemaForm'];
 
-// If ngMock is loaded, it takes over the backend. We should only add
-// it to the list of module dependencies if we are in "frontend mock"
-// mode. Flag this by putting the class .frontendMock on some element
-// in the demo .html page.
-var mockApi = document.querySelector('.mockApi');
-if (mockApi) {
+var angular = require('angular');
+
+// dist/moondash-vendors.js does NOT include ngMockE2E. Only add that
+// dependency and md.mockapi if we have ngMockE2E.
+if (angular.mock) {
   dependencies.push('ngMockE2E');
   dependencies.push('md.mockapi');
 }
 
-var angular = require('angular');
 angular.module('moondash', dependencies);
 
 // Require the Moondash components